refactor : 투자 상품 추천 컴포넌트 분리
